Implement Downloads using POST instead of GET
With the new flag --use-post one can ensure that the FHIR search query
specified with --query is send as POST request in the body.

Using POST can have two benefits, first if the query string is too large
for URL's, it will still fir in the body. Second if the query string
contains sensitive information like IDAT's it will be less likely end up
in log files, because URL's are often logged but bodies not.

The next links are still traversed with GET. The FHIR server is supposed
to not expose any sensitive query params in the URL and also keep the
URL short enough.