Fix Limit of 100 Polls
Async polling had the problem that no more than 100 polls were done
before blazectl got stuck and had to be killed. The problem was caused
by not reading the HTTP responses completely. I did this and moved the
function PollAsyncStatus into the fhir package so that it also can be
used from the compact command.