Improve Download JSON Parsing Performance
The standard encoding/json lib is quite slow because it uses reflection.
There are some third party drop-in replacements like
https://github.com/json-iterator/go or https://github.com/goccy/go-json.
I decided to go with https://github.com/goccy/go-json because it's
actively maintained.

The speedup is about 2x. The download of 1 million cached patient
resources finishes in 11 seconds compared to 25 seconds before. Other
downloads of real FHIR searches or uncached resources is not faster,
but the CPU usage of blazectl is about half as much.